+++ Changelog.md
@@ -4,9 +4,10 @@ Builtin:
  - `name` is now typed as `any -> variadic any prop` to support the following
    two use cases:
    + `name T` succeeds if `T` is an eigenvariable (even if it is applied)
-   + `name T HD ARGS` succeeds if `T` is an eigenvariable an decomposes it
-     to its head symbol and its arguments as a list:
+   + `name T HD ARGS` relates `T` (an applied eigenvariable) to its head
+     and arguments (as a list):
      `pi f x y\ name (f x y) f [x,y]`
+ - new builtin `constant` working as `name` but for non-eigenvariables
 
 ## Version 1.1 (September 2018)